--- a/http/nginx.conf
+++ b/http/nginx.conf
@@ -5,6 +5,9 @@ events {
 }
 
 http {
+    # Hide nginx version (globally)
+    server_tokens off;
+
     server {
         listen 80;
         listen [::]:80;