--- a/http/nginx.conf
+++ b/http/nginx.conf
@@ -32,6 +32,26 @@ http {
         ssl_certificate /etc/nginx/ssl/live/studi-ecf.markomilicevic.fr/fullchain.pem;
         ssl_certificate_key /etc/nginx/ssl/live/studi-ecf.markomilicevic.fr/privkey.pem;
 
+        # Enable Gzip compression (excepted on very old browsers)
+        gzip on;
+        gzip_disable "msie6";
+        gzip_comp_level 6;
+        gzip_min_length 1100;
+        gzip_buffers 16 8k;
+        gzip_proxied any;
+        gzip_types
+            text/plain
+            text/css
+            text/js
+            text/xml
+            text/javascript
+            application/javascript
+            application/x-javascript
+            application/json
+            application/xml
+            application/rss+xml
+            image/svg+xml/javascript;
+
         root /var/www/website/;
 
         try_files $uri $uri/ /index.html =404;