--- a/http/nginx.conf
+++ b/http/nginx.conf
@@ -100,6 +100,8 @@ http {
 
         location /img/ {
             alias /var/www/static/;
+            expires 1y;
+            add_header Cache-Control "public";
         }
 
         location /api/ {
@@ -108,6 +110,11 @@ http {
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
             proxy_set_header X-Forwarded-Proto $scheme;
+        }
+
+        location ~* \.(js|css)$ {
+            expires 1y;
+            add_header Cache-Control "public";
         }
     }
 }