**Here**

Optimize production performances deployment:
- Security: hide nginx version
- Performances: Enable Gzip compression
- Performances: Enable long duration caching for statics, js and css files
- SEO: Mention that the no-index is voluntary

**Tests**

`make run-end2end-tests`

```
 PASS  website/Booking.test.js (10.798 s)
 PASS   website/Contact.test.js (12.869 s)
 PASS  website/Movies.test.js (14.314 s)
 PASS  website/HomePage.test.js (20.56 s)
 PASS  website/Account.test.js (24.518 s)
```

**Demo**

Before:

![gtmetrix com_reports_studi-ecf markomilicevic fr_46jmrhDC_ (1)](https://github.com/markomilicevic/studi-ecf/assets/3176231/dfd23ff6-c6dc-42d3-9346-8ad7bba8045d)