@markomilicevic 2024-06-30
After:

![gtmetrix com_reports_studi-ecf markomilicevic fr_HMMNRBHX_](https://github.com/markomilicevic/studi-ecf/assets/3176231/f29cee42-80bd-4944-9e71-0982197be0aa)